requirements.txt
langchain-google-genai
langchain
langchain-core
langchain-community
chromadb==0.5.23
pypdf
python-dotenv
google-generativeai
